WIP: Eventually consistent host state prototype
This patch is a conceptual implementation of bp
eventually-consistent-scheduler-host-state to testify its design goals
in accuracy, performance, reliability and compatibility improvements.

To run and test this prototype, download this patch and set
`scheduler_host_manager=shared_host_manager` in "nova.conf". Do not
forget to update "nova/nova.egg-info/entry_points.txt" according to the
modified "setup.cfg". Also note: Multiple scheduler instances can only
run on separate machines, and this prototype supports filter schedulers
and caching schedulers running at the same time.

In order to be flexible with design changes and quick prototyping, there
are still no unit tests. They will be added once the general idea is
accepted and the design is confirmed with spec.

For discussions and detailed explanation, please refer to the maillist:
http://lists.openstack.org/pipermail/openstack-dev/2016-February/086516.html

If you find any bugs, please paste your detailed log information to
http://paste.openstack.org/ and reply here.

Please refer to the summit session if you are interested:
https://www.openstack.org/summit/austin-2016/vote-for-speakers/Presentation/7316